requirements.txt
PyYAML>=6.0
python-can>=4.0
pyserial>=3.5
